GUI Refactor. Add progress bar. Remove tqdm and loss extra console lines
