+++ lib/gui/__init__.py
@@ -1,0 +1,4 @@
+from lib.gui.utils import Tooltip, Utils
+from lib.gui.console import ConsoleOut
+from lib.gui.command import CommandTab
+from lib.gui.display import DisplayTab
