+++ lib/gui/icons/clear.png
