+++ lib/gui/icons/open_folder.png
