+++ lib/gui/icons/reset.png
